Restore Python 3 compatibility (issue #2)
This is supposed to fix the test broken in revision 1da3dd8f25b.
In local testing with tox the test indeed works again (without
any changes to the test compared to previous revisions, which
means the sorting order hasn't changed, as intented :-).

For details refer to issue 2 on GitHub:
  https://github.com/xolox/python-naturalsort/issues/2